Move JobHostContext creation logic to JobHostContextFactory (#305).
